Batch ins payment amounts would be off by a penny rarely.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@7465 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormClaimPayBatch.cs
@@ -654,9 +654,9 @@ namespace OpenDental{
 				gridAttached.Rows.Add(row);
 			}
 			gridAttached.EndUpdate();
-			double total=0;
+			decimal total=0;
 			for(int i=0;i<ClaimsAttached.Count;i++) {
-				total+=ClaimsAttached[i].InsPayAmt;
+				total+=(decimal)ClaimsAttached[i].InsPayAmt;
 			}
 			textTotal.Text=total.ToString("F");
 			//gridOutstanding-------------------------------------------------------------------------------------------------
